Add an option to output in ntop's TLV format.
As of ntop v6.4, the JSON format is no longer accepted for
community mode: https://github.com/synfinatic/netflow2ng/issues/89

This commit adds an --tlv commandline parameter that will be
accepted by ntopng as of v6.4

The format is "interesting" so I put it all in its own file. See
that file for details on the format.

TESTING:
* Running netflow2ng on Debian 13, router sending it Netflow v9
* ntopng v6.4 (installed from packages) on different host.
* Confirmed ntop was in community mode and
  that I could see recent flows that matched my network activity.

FUTURE WORK:
* Test against top-of-tree ntopng
* Update the ZmqHeader struct to the most recent one in the ntopng
  code, in an effort to stay ahead of any more deprecation.