Use the latest zmq_msg_hdr_v3 from ntop_typedefs.h
Update the ZMQ header to the latest version in as of ntop 6.4.
The goal here is to just use the latest thing so we have longer
before we have to worry about being deprecated.

The work here was:
* New header with new version constant

* New header requires knowing compress/uncompressed sizes so move
  the compress code from Format to Transport. We're still only
  compressing JSON.

TESTING DONE:
Tested TLV, JSON, and compressed JSON with ntopng 6.4 built from
source. Saw flows with all three. In order to get 6.4 to work with
compressed JSON, I had to fix a bug in ntopng as well as disable
the pro license checks.

Without this fix, I cannot see how ntopng decompresses anything
that has zmq_message_header_v3. I confirmed with them this is
fixed in their dev branch, but broken in v6.4