Make Netflow templates available via HTTP again.
Add back the old-goflow feature to look at NF templates via an
http endpoint. In order do this, I submitted a PR to the goflow2
project. Until that is merged, using a replace line in go.mod to
point to my fork of goflow2 that contains the required changes
for the /templates endpoint to work.

This commit can be merged as is or left until the goflow2 folks
merge my PR.